Dev_0.0.12

+++ TestTask.Users/TestTask.Users/Handlers/Commands/DeleteUserCommandHandler.cs
@@ -1,13 +1,16 @@
 ﻿using AutoMapper;
 using AzureFromTheTrenches.Commanding.Abstractions;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 using TestTask.Users.BLL.DTOs.Users;
 using TestTask.Users.BLL.Services.Contracts;
 using TestTask.Users.Commands;
+using TestTask.Users.Handlers.Abstracts;
 
 namespace TestTask.Users.Handlers.Commands
 {
-    public class DeleteUserCommandHandler : ICommandHandler<DeleteUserCommand>
+    public class DeleteUserCommandHandler : BaseActionHandler<DeleteUserCommand>,
+        ICommandHandler<DeleteUserCommand, IActionResult>
     {
         private readonly IUserService _userService;
 
@@ -15,11 +18,17 @@ namespace TestTask.Users.Handlers.Commands
         {
             _userService = userService;
         }
-        public async Task ExecuteAsync(DeleteUserCommand command)
+
+        protected override async Task<IActionResult> ExecuteAction(DeleteUserCommand command)
         {
             GetUserDTO user = await _userService.GetUserAsync(command.Id);
 
-            await _userService.DeleteUserAsync(user);
+            if (user != null)
+            {
+                await _userService.DeleteUserAsync(user);
+            }
+
+            return Ok(user);
         }
     }
 }
